(#1211) - Uniquify GET requests to bust IE caching
IE10/11 fail in a variety of scenarios where we expect
that the result of an HTTP GET is always fresh,
because IE often aggressively caches AJAX requests.
CouchDB doesn't even get the request, and IE pretends
it got a 304.  Apparently this bug has been around since
2006; there's a good writeup here:
http://www.dashbay.com/2011/05/internet-explorer-caches-ajax/

My solution is to make the ajax library similar to jQuery's;
i.e. there's a param "cache" that, if set to false, adds
a nonce to each GET request. I'd love to make this an IE-only
thing, but apparently IE11 masquerades as Firefox with
its user-agent, so it's better just to do it everywhere.